requirements.txt
amqp==5.3.1
APScheduler==3.10.4
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1711268871457/work
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
biopython==1.78
bleach==6.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
conda-pack @ file:///home/conda/feedstock_root/build_artifacts/conda-pack_1731874358732/work
decorator==5.1.1
defusedxml==0.7.1
Django @ file:///opt/conda/conda-bld/django_1661284230642/work
django-cors-headers==4.3.1
django-redis==5.4.0
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.21.1
idna==3.10
importlib_metadata==8.5.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kombu==5.4.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.26.4
packaging==24.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2 @ file:///croot/psycopg2_1704232443350/work
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz @ file:///croot/pytz_1713974312559/work
pyzmq==26.2.0
redis==5.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.0
six==1.16.0
soupsieve==2.6
sqlparse @ file:///croot/sqlparse_1690901927333/work
stack-data==0.6.3
supervisor==4.2.5
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zipp==3.21.0
zipstream==1.1.4
